fix:naming conventions and inline styles
