+++ App.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 
-import {IsLoggedInProvider} from '@src/contexts/isLoggedInContext';
 import Navigator from '@src/navigation/Navigator';
+import {UserLoginProvider} from '@src/contexts/isLoggedInContext';
 
 function App(): React.JSX.Element {
   return (
     <NavigationContainer>
-      <IsLoggedInProvider>
+      <UserLoginProvider>
         <Navigator />
-      </IsLoggedInProvider>
+      </UserLoginProvider>
     </NavigationContainer>
   );
 }
