--- src/components/bottomAuthScreen/index.ts
@@ -1,3 +1,0 @@
-import BottomAuthScreen from "./BottomAuthScreen";
-
-export default BottomAuthScreen
