--- src/components/bottomAuthScreen/styles.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-    container:{
-        flexDirection: 'row',
-        alignSelf: 'center',
-        marginTop: 4
-    },
-    text:{
-        marginTop: 8
-    }
-})
-
-export default styles;
