+++ src/components/loginForm/index.ts
@@ -1,0 +1,3 @@
+import LoginForm from "./Loginform";
+
+export default LoginForm
