+++ src/components/loginForm/styles.ts
@@ -1,0 +1,21 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  image: {
+    height: 100,
+    width: 350,
+    alignSelf: 'center',
+    marginTop: 60,
+    marginBottom: 20,
+  },
+  buttonContainer: {
+    marginTop: 4,
+  },
+  text: {
+    fontSize: 32,
+    marginHorizontal: 10,
+    textAlign: 'center',
+    color: '#4169E1',
+    marginBottom: 32,
+  },
+});
