+++ src/screens/signUp/SignUp.tsx
@@ -1,14 +1,15 @@
 
 import {ROUTES} from '@src/constants/Routes';
-import BottomAuthScreen from '@src/components/bottomAuthScreen';
+
 import SignUpForm from '@src/components/signUpForm/SignUpForm';
 import BaseContainer from '@src/components/baseContainer';
+import Footer from '@src/components/footer';
 
 const SignUp = ({navigation}: SignupScreenProps) => {
   return (
     <BaseContainer>
       <SignUpForm />
-      <BottomAuthScreen
+      <Footer
         text="Already have an account?"
         buttonLabel="Login"
         onPressLogin={() => navigation.replace(ROUTES.LOGIN)}
