+++ src/theme/ViewStyles.ts
@@ -1,8 +1,8 @@
 import {StyleSheet} from 'react-native';
 
-export const ViewStyles = StyleSheet.create({
-  View: {
-    flex:1,
+export const viewStyles = StyleSheet.create({
+  view: {
+    flex: 1,
     backgroundColor: 'white',
   },
 });
